app/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Redis==0.4.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
marshmallow==3.20.1
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1